Split Behavior config section into focused sections
Replace the single long "Behavior" settings list with four top-level
config sections that plug into the existing accordion/sidebar nav:

- Display: auto-hide scoreboard, auto simple mode
- Statistics: point-type tracking, detailed opponent errors
- Set summary: recap overlay enable, default style, auto-show timing
- General: haptic feedback, keyboard shortcuts, language

Every individual setting keeps its existing behavior and i18n field
keys; only the grouping moved, so options are easier to scan and find.
New section labels are localized across all six locales.

Updates ConfigPanel section registry (type/order/keys/icons) and the
language-selector test (now under "General"). Removes BehaviorSection.
