+++ frontend/src/i18n/translations.ts
@@ -125,7 +125,10 @@ export const translations: Record<string, TranslationDict> = {
     'section.overlay': 'Overlay Style',
     'section.position': 'Position & Size',
     'section.buttons': 'Button Appearance',
-    'section.behavior': 'Behavior',
+    'section.display': 'Display',
+    'section.stats': 'Statistics',
+    'section.recap': 'Set summary',
+    'section.general': 'General',
     'section.rules': 'Match rules',
     'section.links': 'Links',
     'presets.apply': 'Load',
@@ -427,7 +430,10 @@ export const translations: Record<string, TranslationDict> = {
     'section.overlay': 'Estilo del Overlay',
     'section.position': 'Posición y Tamaño',
     'section.buttons': 'Apariencia de Botones',
-    'section.behavior': 'Comportamiento',
+    'section.display': 'Visualización',
+    'section.stats': 'Estadísticas',
+    'section.recap': 'Resumen de set',
+    'section.general': 'General',
     'section.rules': 'Reglas del partido',
     'section.links': 'Enlaces',
     'presets.apply': 'Cargar',
@@ -728,7 +734,10 @@ export const translations: Record<string, TranslationDict> = {
     'section.overlay': 'Estilo do Overlay',
     'section.position': 'Posição e Tamanho',
     'section.buttons': 'Aparência dos Botões',
-    'section.behavior': 'Comportamento',
+    'section.display': 'Visualização',
+    'section.stats': 'Estatísticas',
+    'section.recap': 'Resumo do set',
+    'section.general': 'Geral',
     'section.rules': 'Regras do jogo',
     'section.links': 'Ligações',
     'presets.apply': 'Carregar',
@@ -1029,7 +1038,10 @@ export const translations: Record<string, TranslationDict> = {
     'section.overlay': 'Stile Overlay',
     'section.position': 'Posizione e Dimensione',
     'section.buttons': 'Aspetto dei Pulsanti',
-    'section.behavior': 'Comportamento',
+    'section.display': 'Visualizzazione',
+    'section.stats': 'Statistiche',
+    'section.recap': 'Riepilogo set',
+    'section.general': 'Generale',
     'section.rules': 'Regole del match',
     'section.links': 'Link',
     'presets.apply': 'Carica',
@@ -1330,7 +1342,10 @@ export const translations: Record<string, TranslationDict> = {
     'section.overlay': 'Style de l’overlay',
     'section.position': 'Position et taille',
     'section.buttons': 'Apparence des boutons',
-    'section.behavior': 'Comportement',
+    'section.display': 'Affichage',
+    'section.stats': 'Statistiques',
+    'section.recap': 'Résumé du set',
+    'section.general': 'Général',
     'section.rules': 'Règles du match',
     'section.links': 'Liens',
     'presets.apply': 'Charger',
@@ -1632,7 +1647,10 @@ export const translations: Record<string, TranslationDict> = {
     'section.overlay': 'Overlay-Stil',
     'section.position': 'Position und Größe',
     'section.buttons': 'Button-Aussehen',
-    'section.behavior': 'Verhalten',
+    'section.display': 'Anzeige',
+    'section.stats': 'Statistiken',
+    'section.recap': 'Satz-Zusammenfassung',
+    'section.general': 'Allgemein',
     'section.rules': 'Spielregeln',
     'section.links': 'Links',
     'presets.apply': 'Laden',
